+++ src/test/kotlin/uk/co/ben_gibson/git/link/GitHubTest.kt
@@ -8,17 +8,18 @@ import uk.co.ben_gibson.git.link.git.*
 import uk.co.ben_gibson.git.link.ui.LineSelection
 import uk.co.ben_gibson.git.link.url.UrlOptions
 import uk.co.ben_gibson.git.link.url.UrlOptionsFileAtBranch
-import java.net.URL
 import java.util.stream.Stream
 import uk.co.ben_gibson.git.link.url.UrlOptionsCommit
 import uk.co.ben_gibson.git.link.url.UrlOptionsFileAtCommit
-import uk.co.ben_gibson.git.link.url.factory.TemplatedUrlFactoryProvider
+import uk.co.ben_gibson.git.link.url.factory.TemplatedUrlFactory
+import uk.co.ben_gibson.git.link.url.template.UrlTemplates
+import java.net.URI
 
 class GitHubTest {
 
     companion object {
 
-        private val REMOTE_BASE_URL = URL("https://github.com/my/repo")
+        private val REMOTE_BASE_URL = URI("https://github.com/my/repo")
         private const val BRANCH = "master"
         private val COMMIT = Commit("b032a0707beac9a2f24b1b7d97ee4f7156de182c")
         private val FILE = File("Foo.java", false, "src", false)
@@ -76,7 +77,7 @@ class GitHubTest {
     @ParameterizedTest
     @MethodSource("urlExpectationsProvider")
     fun canGenerateUrl(options: UrlOptions, expectedUrl: String) {
-        val factory = TemplatedUrlFactoryProvider().forHost(GitHub())
+        val factory = TemplatedUrlFactory(UrlTemplates.gitHub())
 
         val url = factory.createUrl(options)